requirements.txt
backports.functools-lru-cache==1.5
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2018.4.16
chardet==3.0.4
cycler==0.10.0
Django==1.11.12
django-filter==1.1.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
idna==2.6
kiwisolver==1.0.1
lxml==4.2.1
matplotlib==2.2.2
numpy==1.14.2
pandas==0.22.0
PyJWT==1.6.1
pyparsing==2.2.0
python-dateutil==2.7.2
pytz==2018.4
requests==2.18.4
six==1.11.0
subprocess32==3.2.7
tushare==1.1.7
urllib3==1.22
